Extract shared rotary fallback helper
